Resolve issues from Scrutinizer
 - https://scrutinizer-ci.com/g/ridibooks/simple-notifier/?branch=master
